// front/src/modules/control-plane/shared/types.ts
export interface DataPair {
  id: number;
  name: string;
}
// ...
export type EditorModeType = 'form' | 'code';